# copyuntil(out::IO, in::IO, delim)