--- a/base/exports.jl
+++ b/base/exports.jl
@@ -857,7 +857,6 @@ export
     readline,
     readlines,
     readuntil,
-    readuntil!,
     redirect_stdio,
     redirect_stderr,
     redirect_stdin,