--- a/base/exports.jl
+++ b/base/exports.jl
@@ -857,6 +857,8 @@ export
     readline,
     readlines,
     readuntil,
+    copyuntil,
+    copyline,
     redirect_stdio,
     redirect_stderr,
     redirect_stdin,