--- a/base/io.jl
+++ b/base/io.jl
@@ -864,7 +864,7 @@ never be decreased.
 function readuntil!(s::IO, buffer::AbstractVector{UInt8}, delim::UInt8)
     n = 0
     @inbounds while true
-        n += _readuntil!(s, @view(buffer[begin+n:end]), delim)
+        n += _readuntil!(s, view(buffer, firstindex(buffer)+n:lastindex(buffer)), delim)
         (buffer[end] == delim || eof(s)) && return n
         resize!(buffer, 2*length(buffer)+1)
     end