--- a/doc/src/base/io-network.md
+++ b/doc/src/base/io-network.md
@@ -69,7 +69,6 @@ Base.dump
 Meta.@dump
 Base.readline
 Base.readuntil
-Base.readuntil!
 Base.readlines
 Base.eachline
 Base.displaysize