This PR defines and exports  ~~a new function `readuntil!(s::IO, buffer::AbstractVector{UInt8}, delim)`~~ new functions:
```jl
copyuntil(out::IO, s::IO, delim; keep=false)
copyline(out::IO, s::IO; keep=false)
```
that read data from `s` into ~~`buffer` in-place (resized if needed)~~ the `out` stream until `delim` is read/written or the end of the stream is reached.

The PR was inspired by [this post](https://discourse.julialang.org/t/is-julia-well-suited-for-string-manipulation/92943/8?u=stevengj) from @jakobnissen: the goal is to make it easier implement an allocation-free `eachline` iterator.   This can be done in a package, given `readuntil` with an `IOBuffer`, using the [`StringViews.jl` package](https://github.com/JuliaStrings/StringViews.jl) to return a string view of the in-place `buffer` on each iteration.

The reason it seemed like this needed a `Base` function, instead of living completely in a package, is that `readuntil` relies on a [low-level `jl_readuntil` C function](https://github.com/JuliaLang/julia/blob/428d242f9b8808e48905c7e82dede0e605f163cb/src/sys.c#L261-L317) that would be difficult to replicate in a package.  To obtain comparable performance, it seems like we need an analogous `jl_readuntil_buf` method (implemented in this PR) and a corresponding Julia API.

Moreover, relatively little new code was required because many of the existing `readuntil` methods used an `IOBuffer` internally, so it was merely a matter of refactoring and exporting this functionality.   Also, we already had an optimized [ios_copyuntil function](https://github.com/JuliaLang/julia/blob/8a9589d5a5d9f6bbc3dd8cbcdfb93fa03527c796/src/support/ios.c#L832-L859) for copying between IOStreams, which can now be exported in the new API.

To do:
- [x] Fix bootstrapping failures
- [x] Benchmarks.  (Maybe it is faster just to read the file in 4k blocks into a buffer with `readbytes!` and then return StringViews on top of that?   This could happen completely in a package.  It's a *lot* easier to use something like `readuntil!`, however.)
- [x] Tests
- [x] More docs
- [x] NEWS
- [x] Fixes and tests for new `out::IO` variant
- [x] add `readline(out::IO, in::IO)` too?
- [x] more tests for `readline(out::IO, in::IO)` methods
- [x] more benchmarks and optimization

Before I do much more work on this, what do people think?