@stevengj 2023-01-14
Yes, they do have an internal buffer, though I'm not sure how large it is.  As I mentioned, the alternative is to do some manual buffering, e.g. reading a file in 4k chunks similar to #42225.  

Of course, the advantage of manual buffering ala #42225 is that it can be implemented in pure Julia code in a package.   On the the other hand, the big *disadvantage* of manual buffering is that it is problematic for non-seekable streams (e.g. pipes) — if you want to read only a few lines, and then do something else with the remainder of the file, the user wants to be able to read from the end of the last line, not from the end of the last buffer chunk.