@stevengj 2023-01-15
I ran some benchmarks on `master` with the current code in this PR.  My benchmark was to open `io = open("HISTORY.md", "r")`, using the `HISTORY.md` file in the julia repo, and count the length of each line:

* **old** — `sum(length, eachline(io))` — `860.583 μs (14091 allocations: 623.17 KiB)`
* **this PR** — using `readuntil!` with a preallocated buffer and `StringViews` to wrap the buffer in a copy-free string: `671.451 μs (0 allocations: 0 bytes)`
* using `readuntil!` with a preallocated buffer and `StringViews` to wrap the buffer in a copy-free string, but *without* the low-level `jl_readline_buf` function: `18.419 ms (0 allocations: 0 bytes)`
* `sum(length, eachline(io))` with a `BufferedInputStream`: `3.781 ms (31347 allocations: 1.97 MiB)`
* `readuntil!` with a preallocated buffer and a `StringView` on a `BufferedInputStream`: 1.144 ms (0 allocations: 0 bytes)

So, only about a 30% speed improvement by eliminating the `String` allocations.  Decent but not overwhelming.  Still satisfying to see `0 allocations`, however.

If this PR is merged, it would be good to update BufferedStreams.jl to add a specialized `readuntil!` method for `BufferedInputStream` that just copies bytes in a big chunk directly from the buffer rather than the fallback of reading bytes one-by-one.   (If you don't use `memcpy`, you lose.)  Hopefully this will improve the speed even further.

-----------
Benchmark code
```jl
using StringViews, BenchmarkTools, BufferedStreams

io = open("HISTORY.md", "r")

@show sum(length, eachline(seekstart(io), keep=true))
@btime sum(length, eachline(seekstart($io), keep=true))

function doit!(io, buf, delim)
    s = 0
    while !eof(io)
        n = readuntil!(io, buf, delim)
        s += length(StringView(@view buf[1:n]))
    end
    return s
end

buf = Array{UInt8}(undef, 1024)
@show doit!(seekstart(io), buf, '\n')
@btime doit!(seekstart($io), $buf, '\n')

bio = BufferedInputStream(seekstart(io))

@show sum(length, eachline((seekstart(bio); bio), keep=true))
@btime sum(length, eachline((seekstart($bio); $bio), keep=true))

@show doit!((seekstart(bio); bio), buf, '\n')
@btime doit!((seekstart($bio); $bio), $buf, '\n')

close(bio)
```