@stevengj 2023-01-15
I added an optimized `readuntil!` for `IOBuffer`, and modified `readuntil(::IOBuffer, ...)` to use it.  It seems to be a net win even for traditional `eachline` usage that allocates a string on each call.  e.g.
```jl
iob = IOBuffer(read("HISTORY.md"))
@btime sum(length, eachline(seekstart($iob), keep=true))
```
previously gave `1.622 ms (23488 allocations: 1.96 MiB)` and now gives `1.162 ms (23484 allocations: 1.96 MiB)`.

If I use an `IOBuffer` with `readuntil!` and a pre-allocated buffer with `StringView` as above, it gives `697.035 μs (0 allocations: 0 bytes)`.