@stevengj 2023-01-16
Other related threads:

* https://discourse.julialang.org/t/bad-performance-of-eachline-on-stdin/55701/34
* https://discourse.julialang.org/t/slow-gzip-streaming-in-julia-but-not-in-python/56346
* https://discourse.julialang.org/t/package-to-read-process-lines-without-new-allocations/92448