@stevengj 2023-01-16
cc @rickbeeloo, author of the https://github.com/rickbeeloo/ViewReader package.