@stevengj 2023-01-18
I was thinking about it some more, and arguably if you want to have maximum performance for an in-place `eachline`-like iteration, then you maybe want to read from the stream in large blocks (containing many linesa) ala BufferedStreams and then *don't copy* at all if possible — just return a `StringView` directly into the buffer.  (Only a little data motion is required when you get to the end of the buffer, and need to move the final line fragment back to the beginning to read more.)   I think this is what ViewReader.jl does?

From that perspective, however, a `readuntil!` API is unnecessary.