@rickbeeloo 2023-01-18
I also agree reading bigger chunks would do better, especially for those reading from HDDs

> Only a little data motion is required when you get to the end of the buffer, and need to move the final line fragment back to the beginning to read more

This works as long as the "to be finished line" is shorter than the currently allocated buffer. If not it might again not find the newline. That would require some extra logic to see if the newline is found yet and otherwise still increase the buffer and extend. Which, in the worst case would give a buffer ~2x the longest line in the file. I didn't really spend time optimizing that, instead, I allocated two buffers (each being the max line length - given by the user) then flip them around like you said and warn if no newline is found after

> Though it may still be more convenient for some applications, especially if you want to read other data in between reading lines

`readuntil!` will still also still be increasingly faster for longer lines (not like in `HISTORY.md`)