@stevengj 2023-02-09
Marking for `triage` since it would be good to get some feedback from core devs on whether this API is desired.