@stevengj 2023-02-10
I was thinking about this some more in connection with #48625, and I'm starting to feel that a better API would be:
```jl
readuntil(out::IO, in::IO, delim)
```
i.e. add an optional `out::IO` argument instead of a buffer.   Advantages:

* Includes functionality of a buffer, since you can pass a preallocated `IOBuffer`.
* Greater functionality because now you can output to arbitrary streams.
* Little new code since much of the code already uses `IOBuffer` internally; just needs to be refactored.
* We already have an optimized [`ios_copyuntil` function](https://github.com/JuliaLang/julia/blob/8a9589d5a5d9f6bbc3dd8cbcdfb93fa03527c796/src/support/ios.c#L832-L859) for copying between `IOStream`s. The new API would let us export this.