@stevengj 2023-02-11
Updated to use the new `readuntil(out::IO, in::IO, delim)` API.   Not tested yet and probably slightly broken, but should give a good sense of what the code will look like.