@fredrikekre 2023-02-12
There is already `write(out::IO, in::IO)`, which is more or less `readuntil(out, in , eof)`. Perhaps this should also be a `write` method?