@stevengj 2023-02-12
I think it's also more discoverable to stick to the convention that you can add an `out::IO` method as the first argument of a "string" function to make it write to a buffer instead of returning a string.  That way you don't have to go hunting around for "what is the buffer equivalent of `foo(...)`", since it's always `foo(io, ...)`.

For example, we already have `join(...)` and `join(io, ...)`, and I think we should also add an `io` argument to `replace` (#48625), and probably to other functions as well.  They can't all be methods of `write` or `print`.